+++ .prettierrc.json
@@ -1,5 +1,5 @@
 {
-  "tabWidth": 4,
+  "tabWidth": 2,
   "semi": true,
   "singleQuote": true,
   "bracketSpacing": true
