+++ src/main.ts
@@ -1,21 +1,20 @@
 import { Plugin } from 'obsidian';
-import { ChooseFolderModal } from './modal';
-
+import ChooseFolderModal from './ChooseFolderModal';
 
 export default class AdvancedNewFilePlugin extends Plugin {
-	async onload() {
-		console.log('loading plugin');
+  async onload() {
+    console.log('loading plugin');
 
-		this.addCommand({
-			id: 'advanced-new-file',
-			name: 'Create note',
-			callback: () => {
-				new ChooseFolderModal(this.app).open();
-			}
-		});
-	}
+    this.addCommand({
+      id: 'advanced-new-file',
+      name: 'Create note',
+      callback: () => {
+        new ChooseFolderModal(this.app).open();
+      },
+    });
+  }
 
-	onunload() {
-		console.log('unloading plugin');
-	}
+  onunload() {
+    console.log('unloading plugin');
+  }
 }
